requirements.txt
langchain
langchain-groq
groq
requests
python-dotenv
